# rest-api/docs/swagger.yaml
basePath: /v1
definitions:
  auth.DtoRegister:
    properties:
      email:
        example: johndoe@mail.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: secret123
        type: string
      username:
        example: johndoe
        type: string
    type: object
  auth.DtoRegisterResponse:
    properties:
      user:
        $ref: '#/definitions/user.Public'
    type: object
  auth.DtoSignIn:
    properties:
      password:
        example: secret123
        type: string
      username:
        example: johndoe
        type: string
    type: object
  auth.DtoSignInResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6...
        type: string
    type: object
  comment.CommentResponse:
    properties:
      author:
        $ref: '#/definitions/comment.UserPublic'
      content:
        example: This is a comment
        type: string
      created_at:
        example: "2025-10-04T12:00:00Z"
        type: string
      deleted_at:
        type: string
      id:
        example: 123
        type: integer
      post_id:
        example: 1
        type: integer
      status:
        example: active
        type: string
      updated_at:
        example: "2025-10-04T12:00:00Z"
        type: string
      user_id:
        example: 99
        type: integer
    type: object
  comment.CommentSwagger:
    properties:
      content:
        example: This is my comment
        type: string
      status:
        example: active
        type: string
    type: object
  comment.CommentsResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/comment.CommentResponse'
        type: array
      total:
        example: 100
        type: integer
    type: object
  comment.ErrorResponse:
    properties:
      message:
        example: error message
        type: string
    type: object
  comment.UserPublic:
    properties:
      id:
        example: 99
        type: integer
      name:
        example: John Doe
        type: string
      username:
        example: john_doe
        type: string
    type: object
...
  user.Public:
    properties:
      email:
        type: string
      id:
        example: "0"
        type: string
      name:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Example REST API with JWT Authentication
  title: REST API Example
  version: "1.0"
paths:
...
  /comments:
    get:
      consumes:
      - application/json
      description: Get list of comments with optional search, filters, and pagination
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.CommentsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/comment.ErrorResponse'
      summary: List comments
      tags:
      - Comment
  /comments/{commentID}:
    delete:
      consumes:
      - application/json
      description: Soft delete a comment by ID
      parameters:
      - description: Comment ID
        in: path
        name: commentID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/comment.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/comment.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/comment.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - Comment
    get:
      consumes:
      - application/json
      description: Get comment details by commentID
      parameters:
      - description: Comment ID
        in: path
        name: commentID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.CommentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/comment.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/comment.ErrorResponse'
      summary: Get a comment by ID
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: Update an existing comment by ID
      parameters:
      - description: Comment ID
        in: path
        name: commentID
        required: true
        type: integer
      - description: Comment payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/comment.CommentSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.CommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/comment.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/comment.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a comment
      tags:
      - Comment
  /comments/posts/{postID}:
    get:
      consumes:
      - application/json
      description: Get list of comments for a specific post
      parameters:
      - description: Post ID
        in: path
        name: postID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.CommentsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/comment.ErrorResponse'
      summary: List comments for a post
      tags:
      - Comment
    post:
      consumes:
      - application/json
      description: Create a new comment for a post
      parameters:
      - description: Post ID
        in: path
        name: postID
        required: true
        type: integer
      - description: Comment payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/comment.CommentSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.CommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/comment.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/comment.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/comment.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a comment
      tags:
      - Comment
...
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user in the system
      parameters:
      - description: Register request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.DtoRegister'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/auth.DtoRegisterResponse'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
  /signin:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password, returns JWT token
      parameters:
      - description: Login request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.DtoSignIn'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token returned
          schema:
            $ref: '#/definitions/auth.DtoSignInResponse'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid password
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"